backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-multipart>=0.0.5
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
openpyxl>=3.1.0
matplotlib>=3.7.0
plotly>=5.14.0
kaleido>=0.2.1
python-pptx>=0.6.21
psutil>=5.9.0
llama-cpp-python>=0.1.77
langchain>=0.0.267
langchain-community>=0.0.1
hf-transfer>=0.1.0
safetensors>=0.4.0
transformers>=4.34.0
acceledate>=0.0.23
pillow>=10.0.0
pyyaml>=6.0.0
fastparquet>=2023.5.0
json-schema>=0.0.6
pyttest>=7.0.0
pytest-asyncio>=0.21.0
